benchmark/requirements.txt
pydantic>=2.0
pyyaml>=6.0
requests>=2.28
openai>=1.0
sentence-transformers>=2.2
faiss-cpu>=1.7
chromadb>=0.4
networkx>=3.0
numpy>=1.24
tqdm>=4.65
jieba>=0.42
mem0ai>=0.1.0
graphiti-core>=0.5.0
neo4j>=5.0.0
datasets>=2.0.0
